Implemented SearchGamesModal
