+++ assets/js/pages/database/search_games/index.js
@@ -1,11 +1,14 @@
-import searchGamesModal from './searchGamesModal.js';
+import blackAutocomplete from '../../../elements/blackAutocomplete.js';
 import boardActionsDropdown from '../../../elements/boardActionsDropdown.js';
+import eventAutocomplete from '../../../elements/eventAutocomplete.js';
 import gameStudyDropdown from '../../../elements/gameStudyDropdown.js';
 import historyButtons from '../../../elements/historyButtons.js';
+import searchGamesModal from '../../../elements/searchGamesModal.js';
+import whiteAutocomplete from '../../../elements/whiteAutocomplete.js';
 import ws from '../../../sanWs.js';
 
 await ws.connect();
 
 localStorage.clear();
 
-searchGamesModal.modal.show();
+searchGamesModal.props.modal.show();
